5.1.14 create Film presenter
